Add the plum script feature. You can create a script like this : ======
setup=(192.168.8.64 00:D0:4B:8B:35:3B)

version
help

======
Note the shebang and the setup= line, these two
things are mandatory if you want to make it work.

You must also put an emty line at the end of your script
otherwise, the last line won't be executed.

The order in the setup does not matter.
See the changelog in debian folder for information on
the -p option, manual is also available.

Signed-off-by: Maxime <maxime.hadjinlian@gmail.com>